Alternative solution to the issue with the getting started tour
