+++ src/panels/GettingStartedPanel/Steps/NavigationSteps.tsx
@@ -1,6 +1,8 @@
-import { Center, Group, Kbd, List, Text, Title } from '@mantine/core';
+import { Button, Center, Group, Kbd, List, Text, Title } from '@mantine/core';
 
-import { OriginPanelMenuButton } from '@/panels/OriginPanel/OriginPanelMenuButton';
+import { FocusIcon } from '@/icons/icons';
+import { FocusButtonContent } from '@/panels/OriginPanel/MenuButtons/FocusButtonContent';
+import { IconSize } from '@/types/enums';
 
 import { AltitudeMouse } from '../MouseDescriptions/AltitudeMouse';
 import { Mouse } from '../MouseDescriptions/Mouse';
@@ -76,7 +78,16 @@ export const NavigationSteps: React.JSX.Element[] = [
     <Text>
       Open the navigation menu and click on the "Focus" button to focus on the Moon. The
       navigation menu can be found in the bottom bar:
-      <OriginPanelMenuButton id={''} />
+      <Button
+        variant={'menubar'}
+        leftSection={<FocusIcon size={IconSize.lg} />}
+        style={{ pointerEvents: 'none' }}
+        px={'sm'}
+        m={'sm'}
+        size={'lg'}
+      >
+        <FocusButtonContent />
+      </Button>
     </Text>
   </>,
   <>
