+++ src/panels/Menu/TaskBar/TaskBarMenuButton.tsx
@@ -14,16 +14,6 @@ export function TaskBarMenuButton({ id, children, ...props }: Props) {
   );
   const { addWindow, closeWindow } = useWindowLayoutProvider();
 
-  if (id === '') {
-    // This is a special case used to show the buttons in the getting started guide.
-    // These buttons should not be clickable
-    return (
-      <Button px={'sm'} size={'lg'} {...props} style={{ pointerEvents: 'none' }}>
-        {children}
-      </Button>
-    );
-  }
-
   const item = menuItemsData[id];
 
   if (!item) {
