--- src/panels/OriginPanel/MenuButtons/AnchorAimButtons.tsx
@@ -1,41 +1,0 @@
-import { Group, Stack, Text } from '@mantine/core';
-
-import { AnchorIcon, TelescopeIcon } from '@/icons/icons';
-import { TaskBarMenuButton } from '@/panels/Menu/TaskBar/TaskBarMenuButton';
-import { IconSize } from '@/types/enums';
-
-interface Props {
-  anchorName: string | undefined;
-  aimName: string | undefined;
-  isOpenSpaceReady: boolean;
-  id: string;
-}
-
-export function AnchorAimButtons({ anchorName, aimName, isOpenSpaceReady, id }: Props) {
-  // TODO: make sure Button has a working label for screen readers since we have mixed
-  // icons, text and other elements inside the button
-  return (
-    <TaskBarMenuButton id={id} disabled={!isOpenSpaceReady}>
-      <Group>
-        <Group gap={5} align={'center'}>
-          <AnchorIcon size={IconSize.md} />
-          <Stack gap={0} maw={130} ta={'start'}>
-            <Text truncate>{anchorName}</Text>
-            <Text fw={500} size={'xs'} c={'dimmed'}>
-              Anchor
-            </Text>
-          </Stack>
-        </Group>
-        <Group gap={5}>
-          <TelescopeIcon size={IconSize.md} />
-          <Stack gap={0} maw={130} ta={'start'}>
-            <Text truncate>{aimName}</Text>
-            <Text fw={500} size={'xs'} c={'dimmed'}>
-              Aim
-            </Text>
-          </Stack>
-        </Group>
-      </Group>
-    </TaskBarMenuButton>
-  );
-}
